--- a/pkg/fileutils/fileutils_linux.go
+++ b/pkg/fileutils/fileutils_linux.go
@@ -1,7 +1,4 @@
-//go:build linux || freebsd
-// +build linux freebsd
-
-package fileutils // import "github.com/docker/docker/pkg/fileutils"
+package fileutils
 
 import (
 	"fmt"